feat: Move serverconfig out of internal
