+++ pkg/server/commands/read_changes.go
@@ -9,9 +9,9 @@ import (
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
-	serverconfig "github.com/openfga/openfga/internal/server/config"
 	"github.com/openfga/openfga/pkg/encoder"
 	"github.com/openfga/openfga/pkg/logger"
+	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	serverErrors "github.com/openfga/openfga/pkg/server/errors"
 	"github.com/openfga/openfga/pkg/storage"
 )
