+++ pkg/server/commands/write_authzmodel.go
@@ -11,8 +11,8 @@ import (
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
-	serverconfig "github.com/openfga/openfga/internal/server/config"
 	"github.com/openfga/openfga/pkg/logger"
+	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	serverErrors "github.com/openfga/openfga/pkg/server/errors"
 	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/typesystem"
